requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.34.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.24.0
secure-smtplib==0.1.1
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
